+++ lib/haml_lint.rb
@@ -21,6 +21,7 @@ require 'haml_lint/file_finder'
 require 'haml_lint/runner'
 require 'haml_lint/utils'
 require 'haml_lint/version'
+require 'haml_lint/version_comparer'
 require 'haml_lint/severity'
 
 # Load all parse tree node classes