+++ lib/haml_lint/utils.rb
@@ -145,6 +145,22 @@ module HamlLint
       count
     end
 
+    # Process ERB, providing some values for for versions to it
+    #
+    # @param content [String] the (usually yaml) content to process
+    # @return [String]
+    def process_erb(content)
+      # Variables for use in the ERB's post-processing
+      rubocop_version = HamlLint::VersionComparer.for_rubocop
+
+      ERB.new(content).result(binding)
+    end
+
+    def insert_after_indentation(code, insert)
+      index = code.index(/\S/)
+      "#{code[0...index]}#{insert}#{code[index..-1]}"
+    end
+
     # Calls a block of code with a modified set of environment variables,
     # restoring them once the code has executed.
     #